refactor: move world.ts back to utils folder
