+++ features/step_definitions/import.ts
@@ -1,5 +1,5 @@
 import * as assert from "assert";
-import { Given, Then } from "../supports/world";
+import { Given, Then } from "../ultils/world";
 
 Given(
   "The csv file {string} with content as below:",
